feat: Trim empty space from bottom of generated receipts
This commit introduces a `trimBottom` function that removes any trailing empty space from the bottom of the generated receipt bitmap.

This is achieved by scanning the bitmap from the last row upwards to find the final row containing non-background pixels and then cropping the image to that height. The initial fixed padding of 5 pixels in the total height calculation has also been removed.